jellyfish_image_classification/requirements.txt
fastapi
uvicorn
pydantic
tensorflow
numpy
pandas
keras
pillow
python-multipart